+++ apps/web/app/influencers/components/index.tsx
@@ -1,0 +1,5 @@
+import Content from './_Content';
+
+export const Influencers = {
+  Content,
+};
